pyproject.toml
[project]
name = "cost-predictor"
version = "0.1.0"
description = "Multi-Model LLM Cost Predictor - Monte Carlo simulation for predicting operational costs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend", "config", "tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "hypothesis>=6.150.2",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
